GPT/requirements.txt
numpy
pandas
scipy
QuantLib-Python
yfinance
matplotlib